moving project name/env to trait
